New features
The Record_Grouping_indent class will combine records with the same title and text (based on an option, which is don't by default).

Added a "title" option to the add_object program which works similar to the "new_parent" record.

Added a "new_parent" option to the add_object.generic program.